Save img to clipboard.
Clipboard is lost when the application closes, so this ended up being a more significant refactor. After the screenshot is cropped to the selection box, the Gio window will close, but the application will continue running in the background until the clipboard is overwritten, at which point it will exit.